Calc. field caching.  Right table join rec. ID's.
The formula of a calc. field was processed for every record - but it only needed to be processed once - so we cache the processed value on first record and then read the cache for every record after that.
The join table process tracked the record ID's for both the left and right table - but only the left table ID's were being used - so the right record ID's are no longer tracked/stored/returned.