Bind variables.
Use the question mark ? within the SQL statement for constant data. You can then specify the bind variable data on the gsSQL custom function OR use the function addBindParameter() if using in app script.  The SQL statement is not compiled, but this just helps to simplify the SELECT statement when used as a custom function within the google sheet.