Bug fixes and Test Case Generator.
Bug 1.  Issue with string/number comparisons.  eg.  'where id = 9'.  As parsed, the value is "9" - but if the table is reading ID as a number, the number 9 in the ID field was not matching.
Bug 2.  When no table definition is specified in gsSQL(), and the table name is assumed to be the SHEET NAME, in the case of a NOT IN (select a from b) - the 'b' table was not being loaded.  Worked for IN case, but 'NOT IN' failed.
Made changes to TestSql.js mostly to enable the ability to generate tests cases directly from the test program and write each case to the current sheet.  There must be a tab representing each table in the sheet in order to work properly.  This is not intended to be used outside of my test sheet.