Bind variable naming change.
References to the bind variable in the SELECT must specifically reference the bind data added.  For example, the first bind data added would be referenced as ?1.  With this change, it is no longer necessary to add a bind data point for every bind ?.  Now if one bind data point is referenced more than once in the WHERE, you can point to the data point needed.