#36.  Changes in table data will trigger gsSQL() to re-run.
Previously, the table ranges were found and then loaded using getValues().  The problems is that changes in the table data would not trigger the custom function (gsSQL) to refresh.  The only time it would is when BIND variable data was used and this linked to a CELL (for example having a START/END date for selecting transaction data that could be updated by the user).
In it simplest use, you could (and still can) just enter the SELECT SQL as the one and only parameter.  This would be parsed and the parsed table names would be loaded from sheets using that name as the TAB name.  This method would then only update when the sheet was initially loaded.  The new syntax requires a little more upfront configuration - but the tradeoff is we always keep SELECT results up to date.  The new syntax is as follows:

gsSQL(statement, tableDefinitions, columnOutput, bindData)

* statement - your SQL SELECT statement.
* tableDefinitions - each table requires two entries.   This can be entered for as many tables that are references in the SELECT.  The first entry is "tableName" (which is the logical table name you reference in the select).  The second is the google range.  Enter this like a normal range you would use in any function.  You MUST include as the first row in the data - column titles.
* columnOutput - true/false - do you want to output a column name in the first row of your result data.
* bindData.  List as many bind data items as used in the select.  The first entry is referenced in the select as ?1, the second as ?2....