#35.  JOIN using expression added.
Not that I have ever used a JOIN this way, but it is now possible to specify an expression for the join.  (when I say now possible, I mean for gsSQL - it was always possible in mySql).
I think that a realistic scenario is when data is embedded in a field that would be used to join in another table.  For example if the customer ID was buried in the field, you can now do something like 

join booksales on substr(booksales.customer_id, 2, 1)