AddDate(), CurDate(), DateDiff() functions.
(previous release) Fixed bug when a function contained a string constant and that constant had a comma, it would not behave correctly.  This version just simplified the logic.
Added support for a (weird) case of HAVING.  Normally it is a comparison used on a group by field, but if group by is not used and having is used in the where - it works in MySql - so it works here too.