Imported CSV with Line Feed char
I had an issue where an imported bank transaction CSV file contained a line feed character in the description field.  When performing a select with calculated fields, the program aborted - EVEN though the field with the line feed character was not in the calculated field itself.  It has to do with how I process calculated fields.  A generated code block is passed to javascript to calculate the value and all data from the current record is included.  The javascript processor would fail.  I have decided to replace any tab, line feed and carriage return with a space - but only for data passed in to perform a calculation - e.g. in my case it was the YEAR() function on the transaction date.